Historical Data

Historical quarterly national accounts and economic statistics data from Haver Analytics. Data is from 1970-03-31 until present.

haver

Format

A data frame with 203 rows and 52 variables:

date

Quarterly.

gdp

Nominal Gross Domestic Product (SAAR, Bil.$).

c

Personal Consumption Expenditures (SAAR, Bil.$).

ch

Real Personal Consumption Expenditures

gdph

Real Gross Domestic Product (SAAR, Bil.Chn.2012$)

jc

Personal Consumption Expenditures: Chain Price Index (SA, 2012=100)

jgdp

Gross Domestic Product: Chain Price Index (SA, 2012=100).

jgf

Federal Govt Consumption & Gross Investment: Chn Price Index(SA, 2012=100)

jgs

State & Local Govt Consumption/Gross Investment: Chn Price Index(SA, 2012=100)

jgse

State & Local Govt Consumption Expenditures: Chain Price Index (SA, 2012=100)

jgsi

State & Local Govt Gross Investment: Chain Price Index (SA, 2012=100)

ptgh

Government Consumption & Investment: Contrib to Real GDP % Change (SAAR, %Pt)

ptgsh

State & Local Govt Consumption/Investment: Contrib to Real GDP % Chg (SAAR, %Pt)

ptgfh

Federal Govt Consumption & Investment: Contrib to Real GDP % Change (SAAR, %Pt)

yptmr

Government Social Benefit Payments to Persons: Medicare (SAAR, Bil.$)

yptmd

Government Social Benefit Payments to Persons: Medicaid (SAAR, Bil.$)

yptu

Govt Transf to Persons: Unemployment Insurance Benefits (SAAR, Bil.$)

gtfp

Government Social Benefit Payments to Persons (SAAR, Bil.$)

ypog

Personal Current Transfer Payments to Government (SAAR, Bil.$)

yptx

Personal Current Taxes (SAAR, Bil.$)

ytpi

Government Tax Receipts on Production & Imports (SAAR, Bil.$)

yctlg

Government Tax Receipts on Corporate Income (SAAR, Bil.$)

g

Government Consumption Expenditures & Gross Investment (SAAR, Bil.$)

grcsi

Contributions for Government Social Insurance (SAAR, Bil.$)

dc

Personal Consumption Expenditures: Implicit Price Deflator (SA, 2012=100)

gf

Federal Government Consumption & Gross Investment (SAAR, Bil.$)

gs

State & Local Government Consumption & Gross Investment (SAAR, Bil.$)

gfh

Real Federal Government Consumption & Gross Investment (SAAR, Bil.Chn.2012$)

gsh

Real State & Local Govt Consumption & Gross Investment (SAAR, Bil.Chn.2012$)

gfrpt

Federal Government Personal Current Tax Receipts (SAAR, Bil.$)

gfrpri

Federal Government Tax Receipts on Production & Imports (SAAR, Bil.$)

gfrcp

Federal Government Tax Receipts on Corporate Income (SAAR, Bil.$)

gfrs

Contributions for Federal Government Social Insurance (SAAR, Bil.$)

gftfp

Federal Government Social Benefit Payments to Persons (SAAR, Bil.$)

gfeg

Federal Grants-in-Aid to State & Local Governments (SAAR, Bil.$)

gsrpt

State & Local Government Personal Current Tax Receipts (SAAR, Bil.$)

gsrpri

State & Local Government Tax Receipts on Production & Imports (SAAR, Bil.$)

gsrcp

State & Local Government Tax Receipts on Corporate Income (SAAR, Bil.$)

gsrs

Contributions for State & Local Government Social Insurance (SAAR, Bil.$)

gstfp

State & Local Government Social Benefit Payments to Persons (SAAR, Bil.$)

gset

State & Local Government Expenditures (SAAR, Bil.$)

gfeghhx

Federal Grants to State & Local Governments: Health & Hospitals (SAAR, Mil.$)

gfeghdx

Federal Grants to State & Local Governments: Medicaid (SAAR, Mil.$)

gfeigx

Federal Investment Grants to State and Local Governments (SAAR, Mil.$)

gfsub

Federal Government Subsidies (SAAR, Bil.$)

gssub

State & Local Government Subsidies (SAAR, Bil.$)

gsub

Government Subsidies (SAAR, Bil.$)

gftfbusx

Fed Transfer Payments/Persons: State Unemployment Insur Benefits (SAAR, Mil.$)

pcw

CPI-W: All Items (SA, 1982-84=100), Bureau of Labor Statistics

gdppothq

Real Potential Gross Domestic Product CBO (SAAR, Bil.Chn.2012$)

gdppotq

Potential Gross Domestic Product CBO (SAAR, Bil.$)

recessq

Quarterly NBER Recession/Expansion: Recession Shading